PRO-5247: Update registry address

+++ src/main/java/uk/gov/hmcts/probate/services/submit/RegistryConfig.java
@@ -24,7 +24,7 @@ public class RegistryConfig {
     }
 
     @Bean
-    public Map<Integer, Registry> registryMap() {
+    public Map<Long, Registry> registryMap() {
         return registries
                 .stream()
                 .collect(Collectors.toMap(Registry::getId, s -> s));
